requirements.txt
einops==0.7.0
sentence-transformers==2.2.2